CMake: remove separate sndfile library tagets
It is not CMake's way. Use BUILD_SHARED_LIBS to select static or
 shared build.

Library name now controlled by CMake. It could be cygsndfile-1.dll on
 CygWin platform, libsndfile-1.dll with MinGW or sndfile.dll on Win32 with
 MSVC.

Python generated DEF file no longer contains library name - it is not
 required anyway.

This is the only way to avoid hard detectible errors when linking SndFile
 library.